# Rewrite performance critical parts of net.box in C