--- a/changelogs/unreleased/gh-6241-net-box-rewrite-in-C.md
+++ b/changelogs/unreleased/gh-6241-net-box-rewrite-in-C.md
@@ -1,0 +1,4 @@
+## feature/core
+
+* Improved net.box performance by up to 70% by rewriting hot code paths in C
+  (gh-6241).